FaceGaze: Initialize FaceLandmarker using model and wasm bytes
This change adds a call to the installFaceGazeAssets API and
subsequently initializes the FaceLandmarker using the model and web
assembly bytes.

This change also adds a test, which is currently hidden behind the
mediapipe build flag, for the new logic.

In a follow-up, we'll check-in the mediapipe JS into
accessibility_common/third_party/mediapipe_task_vision/.

Bug: b:309121742
Change-Id: Ieaa58a1db85b3c1d35324c1368f6dbdd389384fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5184763
Reviewed-by: Katie Dektar <katie@chromium.org>
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1252763}