Use accessibility delegate on menu items for better TalkBack focus
Adds the content description to the toggle switch and radio button so
the menu items are properly described by TalkBack and include the
current state of the switches and buttons.

Removes setting the content description for menu items since it is
handled by the accessibility delegate.


Bug: 321786797
Change-Id: Ifea4426ebfa2dd9ccac34e708b3d4eea87488d65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5230720
Reviewed-by: Ian Wells <iwells@chromium.org>
Commit-Queue: Andrea Gomez <andreaxg@google.com>
Cr-Commit-Position: refs/heads/main@{#1252764}