[infra] Add comments explaining the usage of targets.star node types.
targets.star contains a rather large number of node types due to the
complexity of allowing the suites, mixins and variants to be shared
between starlark and //testing/buildbot. To aid understanding, this
change adds comments to each of the node types that indicates when the
nodes are created, what parent and child relationships it has and when
those edges are traversed.

Bug: 1420012
Change-Id: I6c20d87cef5d82a7e4fb6804942489f544044543
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5235192
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1252775}