[infa] Add a node type for target mappings.
The _TARGET_BINARY node type currently represents 2 things: an entry to
be generated in gn_isolate_map.pyl and a valid reference that can be set
as the binary argument for gtests and isolated script tests. However,
not all such references actually make sense since a "binary" with type
"additional_compile_target" can't be used as the binary for a test. This
separates the 2 responsibilities into separate node types so that the
compatibility is enforced by the existence of nodes rather than
requiring additional code to check.

Bug: 1420012
Change-Id: I6db17d4078f90b362f1f4cf9ca65a20cb425840c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5240872
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Commit-Queue: Struan Shrimpton <sshrimp@google.com>
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#1252808}