[import-attributes] Only run WPT tests relying on `with`  in the attributes virtual suite
Multiple WPT tests related to JSON and CSS modules are failing because
they rely on the `with` import attributes syntax, which is currently
disabled by default.

This patch marks them to only be run as part of the
`import-attributes-json-css-modules` virtual test suite, which enables
the `--harmony-import-attributes` flag, so that unflagging the syntax in
V8 (done in Ifc5af7f33aabfa0a2d5cd6c8381561848adbe392, blocked on this
patch) will not cause the tests to suddenly pass breaking the current
test expectations.

Note that for all the tests in
`wpt/html/semantics/scripting-1/the-script-element/css-module` and in
`wpt/html/semantics/scripting-1/the-script-element/json-module` there
are equivalent tests in the `css-module-assertions` and
`json-module-assertions` folders, using the `assert` syntax which is
currently enabled by default. The new expected failures for these tests
in the virtual test suite match the expected failures for the
`assert`-based tests.

Bug: v8:13856
Change-Id: I8a002c10e1091eabd13c15ecaa8e75df6a9ef5e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5237319
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Nicolò Ribaudo <nribaudo@igalia.com>
Reviewed-by: Shu-yu Guo <syg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1252815}