Status bar updates when suggestions shown.
There are two issues fixed here. The first is what when the omnibox
is focused in incognito mode, the color used was never updated for
the modernize visual update change. A new color was simply added.

However the modernize visual update also adjusted the toolbar tint
slightly when the suggestions were scrolled. To solve this the
StatusBarColorController needed to be an observer of this signal and
be wired up to subscribe to the right thing.

Bug: 1519974
Change-Id: Idbc83451f49bcbc58ffe681c645067b3373e67d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5239272
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: Patrick Noland <pnoland@chromium.org>
Commit-Queue: Sky Malice <skym@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1252828}