[Fast Pair][Software Scanning] HardwareOffloadingSupportedProvider
Create the HardwareOffloadingSupportedProvider class. This provider
retrieves a pointer to an adapter object and determines whether
the device supports offloading bluetooth scanning onto its hardware.

Note that the HardwareOffloadingSupportedProvider ctor assumes that
it will receive an adapter object from BluetoothAdapterFactory::Get().
Comments above this function note that it will not return without an
initialized adapter object; for this reason, the circumstance where
the received pointer to the adapter is null is not tested, though
the pointer's validity is still tested before accessing member functions
in the HardwareOffloadingSupportedProvider.

The FakeBluetoothAdapter class is changed to better mock the
BluetoothAdapter class: on a change to status as 'present', it
notifies its observer of this change. A redundant function,
NotifyPoweredChanged, is deleted as well.

        this, pass.

Tested: all unit tests in ash_unittests, including those written in
Fixed: b/322223967
Change-Id: I25e57723024af358988ae01e7f1b452d7e138e76
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5235521
Commit-Queue: Brando Socarras <brandosocarras@google.com>
Reviewed-by: Jack Shira <jackshira@google.com>
Cr-Commit-Position: refs/heads/main@{#1252849}