media/gpu/iptest: Add mode to force a particular IP backend
This is useful for locally testing the VAAPI and V4L2 image processors.
It is not used by any tast test, but could be.

If the backend is forced, then the Vulkan and MM21 comparison tests are
skipped.

Also, if the backend is forced, but the forced backend does not support
the test's port configuration, then the test is skipped. This happens by
checking the return value of the ImageProcessorBackend Create function.
If it returns nullptr for a forced backend, then the test is skipped.

Bug: None
Test: Tests pass without the "force backend" arguments.
Test: Tests produce results with the backend forced.

Change-Id: I3bbaba5f69505861991fecbbc62321db0e65b9d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5152718
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>
Commit-Queue: Nathan Hebert <nhebert@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1252900}