[CodeHealth] Remove dangling pointer in WebContentsModalDialogManager
Clean up the delegate in a few places by setting it to nullptr before
destruction. The locations needing cleanup were identified by a
combination of bot failures after removing the annotation and code
search for callers of SetDelegate() that called this only once and did
not call SetDelegate() again with a nullptr argument.

Bug: 1512712
Change-Id: I35be1e5f139ad28dba8bb3f364c5c753cffc1aa6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5142391
Reviewed-by: Mike Wittman <wittman@chromium.org>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1252985}