Stop using process privilege buckets for extensions.
These are mostly unnecessary now that all extensions require locked
processes, and their remaining use cases can be replaced with
alternate checks. This CL moves their use behind a kill switch feature
(StopUsingRenderProcessHostPrivilege). Namely,
RenderProcessHostPrivilege, GetProcessPrivilege() and
GetPrivilegeRequiredByUrl() all become dead code and can be removed
once this CL is verified not to cause problems in the wild.

The buckets, defined by the RenderProcessHostPrivilege enum, were used
in two places:

1. In ChromeContentBrowserClientExtensionsPart::CanCommitURL(), which
is checked at DidCommitNavigation time to verify that a navigation
commit was valid (terminating the renderer otherwise). There's a
workaround that allows an extension URL to commit in any
PRIV_EXTENSION process, even if that process's ProcessMap entry
doesn't match the committing extension ID (e.g., extension A is
allowed to commit in extension B's process). This was for an old bug
(crbug.com/840857) that predated strict site isolation and strict
extension isolation, and it's highly likely that this is no longer
needed. A DumpWithoutCrashing was added to confirm this in r1239949,
and so far there haven't been any DWOC reports in one month since it
landed (see crbug.com/840857 for details). This CL turns this
workaround off (behind a kill switch).

2. In ChromeContentBrowserClientExtensionsPart::IsSuitableHost(),
which is checked when picking a SiteInstance and a process for a
navigation, to ensure that a navigation can only commit in a process
with a matching privilege bucket. This is generally a less strict
check than site isolation, since it only ensures separation between
the three privilege buckets (all extensions, all hosted apps, all web
sites), but does not check process reuse within each bucket, unlike
core site isolation checks that compare actual process locks (e.g., in
RenderProcessHostImpl::IsSuitableHost(), just prior to calling
ChromeContentBrowserClientExtensionsPart::IsSuitableHost()). However,
the privilege bucket check does have subtle use cases not covered by
core site isolation checks:

(i) It does not allow a hosted app loaded from https://foo.com to reuse
a process of a regular web page loaded from https://foo.com (e.g.,
when one URL is in the hosted app's extent but a second same-site URL
is not, or when reloading a tab after disabling or re-enabling the
hosted app). There are various tests covering this, including
ProcessManagementTest.ProcessOverflow,
AppApiTest.NavigateIntoAppProcess, AppApiTest.ReloadIntoAppProcess*,
and other tests in AppApiTest.

(ii) It does not allow an extension URL loaded for a disabled extension
to share a process with an extension URL loaded after re-enabling the
extension. This is covered by
DisableExtensionBrowserTest.VisitReenabledExtensionInSubframe.

This CL proposes a different way to do these checks, without relying
on process privilege buckets.  Namely, the extensions ProcessMap
already tracks whether a given process is used for particular
extensions or hosted apps, and processes that load disabled
extensions/app URLs are *not* entered into the ProcessMap. So, both
(i) and (ii) above can be handled by checking that (1) a
chrome-extension://extid/ URL can only commit in a process P that's
registered in the ProcessMap for that extension (i.e., has an entry
{extid, P}), and (2) a non-extension URL cannot use a process that's
registered in the ProcessMap for any extension.

This CL replaces the IsSuitableHost privilege bucket checks with these
two new checks, keeping the old path around behind a kill switch as a
precaution. This approach more closely matches the existing checks [1]
in CanCommitURL(), which are already based on ProcessMap. Since
IsSuitableHost() is provided with a target site URL that's computed
after resolving effective URLs, this works for both extensions and
hosted apps. A URL loaded in a hosted app won't be allowed to share a
process with a same-site non-app URL, since the former will get
resolved to a chrome-extension: site URL, with a process that's
registered in ProcessMap, while the latter won't have a
chrome-extension: site URL, with a process that's not registered in
ProcessMap.

This CL builds on creis@'s earlier work in
https://chromium-review.googlesource.com/c/chromium/src/+/4127610,
taking a slightly stricter approach of using ProcessMap-based checks
instead of checking that the extension process should be locked (which
should already be checked earlier in RPHI::IsSuitableHost()). This
also handles the disabled extension cases and allows us to remove the
hosted app bucket.

There is one known behavior difference caused by the new ProcessMap
checks in this CL. Previously, two hosted apps loaded from different
parts of the same site (e.g., from https://app.com/app1/app1.html and
https://app.com/app2/app2.html) were allowed to share a process (e.g.,
when over the process limit), since they ended up with the same
process lock (https://app.com) and the same privilege bucket
(PRIV_HOSTED). The new ProcessMap checks are stricter and only allow
an extension X to use a process registered in the ProcessMap
specifically for X and not another extension. Hence, the process reuse
in the above scenario is no longer allowed - only instances of the
same app are allowed to share a process. This behavior should be more
secure and shouldn't actually cause any breakage, since we wouldn't
normally reuse a process for these apps unless we were over the
process limit. This is covered by one of the cases in
ProcessManagementTest.ProcessOverflow.

[1] https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/extensions/chrome_content_browser_client_extensions_part.cc;l=429;drc=b61931b0d79ccc8bfa081f001cd2b53b1ea98107

Change-Id: I24fd96727221ed679ca36703ef0121ab0c27531c
Bug: 1519931, 840857
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5187071
Commit-Queue: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1253001}