Don't do sync about:blank nav for iframe navs not initiated by parent
It's possible for about:blank navigations to happen on an iframe that's
currently on the initial empty document, but the initiator origin is
not the parent's origin, e.g. if it's started by another ancestor frame
or child frames. In this case, don't categorize the navigations as
synchronous about:blank navigations.

Bug: 1476614
Change-Id: Ic8b1af40112ca564305802066c80841824a00a80
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4880842
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1253106}