Remove CrashWithoutDump() for missing ColorProvider in AppControllerMac
During the browser startup the creation of Browser and AppController is
a race condition. The color provider will be missing if the browser is
created later than the AppController.

This CL removes the CrashWithoutDump() intended to study this triggering
condition and adds comment to explain.

Bug: 1522363
Change-Id: I6930b813495590888a6ee51e7c35c01d979baf76
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5240988
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Keren Zhu <kerenzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1253112}