Implement sanitization feature for ChromeOS
This CL implements the sanitization feature for ChromeOS. This feature
aims to provide a "soft reset" to the users. In this feature some
settings which could potentially be used for malicious purposes are
disabled/reset. For example all extensions are disabled, and homepage
settings is reset. However, cookies are not removed and no personal data
is removed as part of this process. This CL is the first of N CLs
needed for this feature but currently includes most of the
functionality. The rest of the CLs will focus more on user experience.
The screenshots for the sanitize row and dialog could be found in the
following links:
https://screenshot.googleplex.com/3S7WL7sbRXutbWf.png
https://screenshot.googleplex.com/8zStJQFrok3hAvy.png

Bug: b:317898420, b:317898366
Change-Id: I5f6bae28860eed8da401ab401edbccd01f59de1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4901483
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Commit-Queue: Behnood Momenzadeh <behnoodm@chromium.org>
Reviewed-by: Wes Okuhara <wesokuhara@google.com>
Cr-Commit-Position: refs/heads/main@{#1253154}