[Launcher metrics] Refactor SearchSessionConclusionToString.
Move method to a more generic location. This will be helpful for an
upcoming metrics CL.

Note that within FederatedMetricsManager, the string versions of the
logged SearchSessionConclusion instances will be different as follows:

  quit        -> Quit
  launch      -> Launch
  answer_card -> AnswerCardSeen

This is not an issue because:

a) This information is not currently being used anywhere.
b) If this information is put into use, it is easy to account for the
variations via server-side pre-processing.

Test: existing
Bug: b/319190264, b/262611120
Change-Id: I2a36ce734bea1a466028f70e48d727f566fcb467
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5223688
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Commit-Queue: Amanda Deacon <amandadeacon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1253155}