[LCPP] Remove unnecessary if conditions to call SetLCPPHint()
This is a refactroring CL to simplify the condition when/what to set
LCPP hint data.

|lcpp_hint| is originally set by `MaybeSetLCPPNavigationHint()` in the
LoadingPredictorTabHelper class, from the browser process. This only set
hint data when LCPP features are enabled, or `IsInOutermostMainFrame()`,
or `IsSameDocument()`. Otherwise |lcpp_hint| is nullptr. So, we don't
have to check `IsOutermostMainFrame()` in
`RenderFrameImpl::CommitNavigationWithParams()`.

I'm also skeptical if we should check `lcpp_enable_perf_improvements`
param here. If the param returns true, this doesn't set LCPP data.
I believe this is not intentional.

Bug: 1419756
Change-Id: Ia444622e13ece96ffe68ecea8a23f6cf8b5082a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5240448
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Reviewed-by: Minoru Chikamune <chikamune@chromium.org>
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1253176}