Update OCR service initialization in AxMediaApp.
Service state query and initialization are simplified.

The unittest for `IsOcrServiceEnabled` function is removed as the
function is now simplified to return a boolean, and using the former
approach in unittests to set it is now not possible.
There are browsertests in ScreenAIServiceRouter that ensure OCR service
readiness is correctly sent through this approach.

AX-Relnotes: n/a
Bug: 1520424
Change-Id: I155dac0369a3f48b5bf698be06b5f2241d968998
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5234144
Reviewed-by: David Tseng <dtseng@chromium.org>
Commit-Queue: Ramin Halavati <rhalavati@chromium.org>
Auto-Submit: Ramin Halavati <rhalavati@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1253190}