[iOS][PageInfo] Add content to security subpage
This CL follows crrev.com/c/5185620, and it actually adds the contents
to the security subpage in the revamped page info with the connection
status description and a learn more button.

UX is still deciding what should be the proper alignment of the security
icon in the subpage, if it should be centered within the cell, if it
should be top-aligned, or anything else. For now, we will leave the icon
centered as this is the default behaviour.

The green lock icon when a page is secure has been removed in
crrev.com/c/4672684 when Chrome substituted the page info's lock icon in
the entry point with the tune icon. We have confirmed with UX[1] that
the green lock icon should have not been removed from the security
section, which matches Desktop and Clank implementations. As such, we
will reintroduce the lock green icon with our changes.

Video navigating to the security subpage: http://dr/file/d/1SQ6B1hBeiihv3eY64gVNrxKKMpmvn_4y/view

[1] https://www.figma.com/file/FuNonyqsJOfGXyznOLg6TF/%F0%9F%93%B1Chrome-iOS-(Bling)-Library?type=design&node-id=2939-24297&mode=design&t=ful20xZ4XWXCXsrB-0

Bug: 1512580
Fixed: b:321266043
Change-Id: I2909d7b071ea46444062a93f4d2a9a21d3a2a9d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5214353
Reviewed-by: Ewann Pellé <ewannpv@chromium.org>
Commit-Queue: Filipa Senra <fsenra@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1253211}