[iOS] Perform a more robust check for password fields
The `isOnPasswordField` check was only checking if a field was
obfuscated to confirm that a field is a password field. This
causes websites which obfuscate other fields (such as credit card
number fields) to get incorrect or missing suggestions for those
fields. This CL adds an extra verification that if a field is
confirmed by field prediction to not be a password field, then
we don't interpret this field as a password field.

Bug: 1518018

Change-Id: I206c9d2c8d3c5091ec2c932ba68eefa441946bde
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5235339
Reviewed-by: Maria Kazinova <kazinova@google.com>
Cr-Commit-Position: refs/heads/main@{#1253261}