[Autofill] Remove AutofillAllowDuplicateFormSubmissions, AutofillProbableFormSubmissionInBrowser
The intention behind those features was to move logic from the renderer
to the browser process. The concrete motivation was that it was
believed to be needed for BFcache. That turned out wrong, and the
features were never launched.

Since then, a lot has happened in the renderer. So before rolling out
these features, we'd have to very carefully review if and how to
update the code behind the feature flags.

Since we haven't had the resources to do that in a few years, this CL
removes the code.

Bug: 1117451
Change-Id: I2956ff0f03dc432a0c88e3a451e93973721b8a0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5237014
Commit-Queue: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Auto-Submit: Christoph Schwering <schwering@google.com>
Reviewed-by: Dominic Battre <battre@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1253268}