Add controller for the default store changed bubble
This CL introduces a controller for the new bubble that will be used
to inform the user that their psw store changed.
Additionally this CL introduces a new method that opens pwm settings
page with a query to highlight the store toggle.

Low-Coverage-Reason: TRIVIAL_CHANGE
Bug: 1503146
Change-Id: I1726455efd690866df5b19e908ef7a92e6307943
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5224623
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Karol Sygiet <sygiet@google.com>
Reviewed-by: Viktor Semeniuk <vsemeniuk@google.com>
Cr-Commit-Position: refs/heads/main@{#1253272}